refactored box_decoration.dart into decoration.dart to move more look and feel there The buttons in our app are now controlled through that file. Overall changes to look and feel can be experimented with there. The "LEAVE" button is separate as it has a different look and feel .
Scene Chat is in - "Global" chat exists, but is removed and can no longer be accessed.

fixed a bug in update setting related to the persistent profile in EmailStorage.instance.userData. <-- maintaining that state is...best I think still, but I can see how it can leave certain things out of date. I will continue to think about this.